+++ pkg/resource/deploy/step_generator.go
@@ -745,13 +745,13 @@ func (sg *stepGenerator) generateSteps(event RegisterResourceEvent) ([]Step, boo
 				sg.refreshStates[old] = state
 				sg.refreshAliasLock.Unlock()
 			}
-			contract.AssertNoErrorf(err, "expected a result from refresh step")
 			sg.events <- &continueResourceRefreshEvent{
 				RegisterResourceEvent: event,
 				urn:                   urn,
 				old:                   state,
 				new:                   new,
 				invalid:               invalid,
+				err:                   err,
 			}
 		}()
 
@@ -822,6 +822,10 @@ func (sg *stepGenerator) continueStepsFromRefresh(event ContinueResourceRefreshE
 	urn := event.URN()
 	old := event.Old()
 	new := event.New()
+	err := event.Error()
+	if err != nil {
+		return nil, false, err
+	}
 
 	// If this is a refresh deployment we're _always_ going to do a skip create or refresh step here for
 	// custom non-provider resources. We also need to skip refreshes for custom provider resources if they
@@ -1761,8 +1765,16 @@ func (sg *stepGenerator) continueStepsFromDiff(diffEvent ContinueResourceDiffEve
 			// If the goal state specified an ID, issue an error: the replacement will change the ID, and is
 			// therefore incompatible with the goal state.
 			if goal.ID != "" {
-				const message = "previously-imported resources that still specify an ID may not be replaced; " +
-					"please remove the `import` declaration from your program"
+				replaceDiff := strings.Join(
+					slice.Map(diff.ReplaceKeys, func(k resource.PropertyKey) string {
+						old := old.Inputs[k].String()
+						new := new.Inputs[k].String()
+						return fmt.Sprintf("%s: %s => %s", k, old, new)
+					}),
+					"\n")
+
+				message := "previously-imported resources that still specify an ID may not be replaced; " +
+					"please remove the `import` declaration from your program;\n" + replaceDiff
 				if sg.deployment.opts.DryRun {
 					sg.deployment.ctx.Diag.Warningf(diag.StreamMessage(urn, message, 0))
 				} else {