+++ sdk/python/lib/pulumi/runtime/_callbacks.py
@@ -17,14 +17,14 @@ from __future__ import annotations
 import asyncio
 import traceback
 import uuid
+from collections.abc import Awaitable, Callable, Mapping
 from typing import (
     TYPE_CHECKING,
+    Any,
     Optional,
     Union,
     cast,
 )
-from collections.abc import Callable
-from collections.abc import Awaitable, Mapping
 
 import grpc
 from google.protobuf.message import Message
@@ -40,7 +40,12 @@ from .proto import (
     resource_pb2,
     resource_pb2_grpc,
 )
-from .rpc import deserialize_properties, serialize_properties
+from .rpc import (
+    deserialize_properties,
+    is_rpc_secret,
+    serialize_properties,
+    unwrap_rpc_secret,
+)
 
 if TYPE_CHECKING:
     from ..resource import (
@@ -296,33 +301,53 @@ class _CallbackServicer(callback_pb2_grpc.CallbacksServicer):
                 from ..resource_hooks import ResourceHookArgs
 
                 new_inputs = (
-                    deserialize_properties(
-                        request.new_inputs,
-                        keep_unknowns=True,
+                    outputtify_secrets(
+                        cast(
+                            dict[str, Any],
+                            deserialize_properties(
+                                request.new_inputs,
+                                keep_unknowns=True,
+                            ),
+                        )
                     )
                     if request.new_inputs
                     else None
                 )
                 old_inputs = (
-                    deserialize_properties(
-                        request.old_inputs,
-                        keep_unknowns=True,
+                    outputtify_secrets(
+                        cast(
+                            dict[str, Any],
+                            deserialize_properties(
+                                request.old_inputs,
+                                keep_unknowns=True,
+                            ),
+                        )
                     )
                     if request.old_inputs
                     else None
                 )
                 new_outputs = (
-                    deserialize_properties(
-                        request.new_outputs,
-                        keep_unknowns=True,
+                    outputtify_secrets(
+                        cast(
+                            dict[str, Any],
+                            deserialize_properties(
+                                request.new_outputs,
+                                keep_unknowns=True,
+                            ),
+                        )
                     )
                     if request.new_outputs
                     else None
                 )
                 old_outputs = (
-                    deserialize_properties(
-                        request.old_outputs,
-                        keep_unknowns=True,
+                    outputtify_secrets(
+                        cast(
+                            dict[str, Any],
+                            deserialize_properties(
+                                request.old_outputs,
+                                keep_unknowns=True,
+                            ),
+                        )
                     )
                     if request.old_outputs
                     else None
@@ -609,3 +634,13 @@ class _CallbackServicer(callback_pb2_grpc.CallbacksServicer):
             result.provider = await _create_provider_ref(opts.provider)
 
         return result
+
+
+def outputtify_secrets(props: dict[str, Any]) -> dict[str, Any]:
+    from ..output import Output
+
+    for k, v in props.items():
+        if is_rpc_secret(v):
+            props[k] = Output.secret(unwrap_rpc_secret(v))
+
+    return props